+++ packages/components/pagination/stories/Pagination.stories.tsx
@@ -10,28 +10,46 @@ export default {
 } as Meta;
 
 export const Default = (args: PaginationProps) => {
-  const { activePage, onPageChange, ...otherProps } = args;
+  const {
+    activePage,
+    onPageChange,
+    itemsPerPage,
+    onViewPerPageChange,
+    ...otherProps
+  } = args;
   const [page, setPage] = useState(activePage);
+  const [view, setView] = useState(itemsPerPage);
   useEffect(() => {
     setPage(activePage);
-  }, [activePage]);
+    setView(itemsPerPage);
+  }, [activePage, itemsPerPage]);
 
   const handlePageChange = (p) => {
     onPageChange && onPageChange(p);
     setPage(p);
   };
 
+  const handleViewPerPageChange = (i) => {
+    onViewPerPageChange && onViewPerPageChange(i);
+    setPage(Math.floor((view * page + 1) / i));
+    setView(i);
+  };
+
   return (
-    <Pagination
-      activePage={page}
-      onPageChange={handlePageChange}
-      {...otherProps}
-    />
+    <div style={{ width: '920px' }}>
+      <Pagination
+        activePage={page}
+        onPageChange={handlePageChange}
+        onViewPerPageChange={handleViewPerPageChange}
+        itemsPerPage={view}
+        {...otherProps}
+      />
+    </div>
   );
 };
 
 Default.args = {
-  activePage: 1,
-  totalItems: 120,
-  itemsPerPage: 10,
+  activePage: 0,
+  itemsPerPage: 50,
+  viewPerPageOptions: [20, 50, 100],
 };