+++ packages/forma-36-codemod/utils/deleteProperty.js
@@ -3,6 +3,6 @@ module.exports.deleteProperty = function deleteProperty(
   { propertyName },
 ) {
   return attributes.filter((attribute) => {
-    return attribute.name.name !== propertyName;
+    return attribute.name?.name !== propertyName;
   });
 };